refactor: move clients commands, helper, and tests to cli
