+++ packages/cli/src/commands/clients/rotate.ts
@@ -1,7 +1,7 @@
 import color from '@heroku-cli/color'
 import {Command, flags} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
-import {CliUx} from '@oclif/core'
+import {Args, ux} from '@oclif/core'
 
 export default class ClientsRotate extends Command {
   static description = 'rotate OAuth client secret'
@@ -11,27 +11,29 @@ export default class ClientsRotate extends Command {
     shell: flags.boolean({char: 's', description: 'output in shell format'}),
   }
 
-  static args = [{name: 'id', required: true}]
+  static args = {
+    id: Args.string({required: true}),
+  }
 
   async run() {
     const {args, flags} = await this.parse(ClientsRotate)
 
-    CliUx.ux.action.start(`Updating ${color.cyan(args.id)}`)
+    ux.action.start(`Updating ${color.cyan(args.id)}`)
 
     const {body: client} = await this.heroku.post<Heroku.OAuthClient>(
       `/oauth/clients/${encodeURIComponent(args.id)}/actions/rotate-credentials`,
     )
 
-    CliUx.ux.action.stop()
+    ux.action.stop()
 
     if (flags.json) {
-      CliUx.ux.styledJSON(client)
+      ux.styledJSON(client)
     } else if (flags.shell) {
-      CliUx.ux.log(`HEROKU_OAUTH_ID=${client.id}`)
-      CliUx.ux.log(`HEROKU_OAUTH_SECRET=${client.secret}`)
+      ux.log(`HEROKU_OAUTH_ID=${client.id}`)
+      ux.log(`HEROKU_OAUTH_SECRET=${client.secret}`)
     } else {
-      CliUx.ux.styledHeader(`${client.name}`)
-      CliUx.ux.styledObject(client)
+      ux.styledHeader(`${client.name}`)
+      ux.styledObject(client)
     }
   }
 }
