+++ packages/cli/test/unit/commands/clients/destroy.unit.test.ts
