+++ packages/cli/test/unit/commands/clients/info.unit.test.ts
