+++ packages/cli/test/unit/commands/clients/rotate.unit.test.ts
